Glass_Identification_Prediction
# Glass Identification
Project Description
The dataset describes the chemical properties of glass and involves classifying samples of glass using their chemical properties as one of six classes. The dataset was credited to Vina Spiehler in 1987. The study of classification of types of glass was motivated by criminological investigation. At the scene of the crime, the glass left can be used as evidence...if it is correctly identified!

The chemical compositions are measured as the weight percent in corresponding oxide.
Attribute Information-
1. Id number: 1 to 214
2. RI: refractive index
3.Na: Sodium (unit measurement: weight percent in corresponding oxide, as are attributes 4-10)
4.Mg: Magnesium
5.Al: Aluminum
6.Si: Silicon
7.K: Potassium
8.Ca: Calcium
9.Ba: Barium
10.Fe: Iron
11.Type of glass: (class attribute)
1- building_windows_float_processed
2- building_windows_non_float_processed
3- vehicle_windows_float_processed
4- vehicle_windows_non_float_processed (none in this database)
5- containers
6- tableware
7- headlamps
There are 214 observations in the dataset. The dataset can be divided into window glass (classes 1-4) and non-window glass (classes 5-7). 
Predict : Type of glass

Dataset Link-
https://raw.githubusercontent.com/dsrscientist/dataset3/refs/heads/main/glass.csv